Just use KBV in the base context
Previously we held the position that these "definitional" datasets
should rely on public vocabularies, and not our application vocabulary.

This is untenable given our needs for various specializations and
occasional invention (as evident in the use of "cherry-picked" kbv terms
already present since day one in this base context (Nationality being
the first IIRC)). It also make the "context swapping" in the maintenance
of these datasets very hard to understand (as it's never even been
documented). It seems more manageable to define a core vocab dataset
(kbv, marc et al) and let the regular ones be based directly on that.

This commit is just a preparatory step towards that. What it does allow
is for us to use all unique KBV terms as is in the datasets, without
requiring us to either define and use "public equivalencies", or add
them one by one to this base.