Add ISO639 datatypes and use in languages dataset
Fixes use of the datatypes as properties, and uses datatyped literals
consequently in the table mapping and construct query.

Note that this changes the public language data a bit by renaming the
specific properties involved in the fix.

Also note that the datatypes are consequently in the KBV namespace. This
is mostly since two sub-datatypes of dc:ISO639-2 (for B and T) are not
defined in DC Terms. (We may finally replace the mix of namespaces
within the unified KBV namespace here, but that is a larger change.)