Overcome backwards xl-id form by handling both
This adds the "normal" form (used in XL) as a sameAs to generated
records if they are to have the backwards form mistakenly minted by this
repository.

An upper timestamp per dataset is used to check for if its member
records are to have the backwards form.

(Eventually we want to "garbage collect" this backwards form from XL, to
ensure they don't "squat" on XL id:s in the future.)